+++ scripts/module.js
@@ -21,6 +21,7 @@ import { setupDisplayWeaponDamage } from "./lib/showBaseDamage.js";
 import { setupHideHeaderButtonText } from "./lib/hideHeaderButtonText.js";
 import { setupPauseReplacement } from "./lib/replacePauseInfo.js";
 import { setupStartOfSession } from "./lib/startOfSession.js";
+import { setuplanguageHandling } from "./lib/languageHandling.js";
 
 export const MODULE_ID = "sundry";
 
@@ -28,7 +29,7 @@ Hooks.once("init", async function () {
   setupSettings();
   registerKeybindings();
   loadAllTemplates();
-  versionMigration()
+  versionMigration();
 });
 
 Hooks.once("ready", async function () {
@@ -48,6 +49,8 @@ Hooks.once("ready", async function () {
     getSetting("highlight.pf2e-toolbelt.target-helper.roll")
   );
 
+  setuplanguageHandling(getSetting("highlight.languages-known"));
+
   //Hide
   if (getSetting("hide.sell-all-treasure")) hideSellAllTreasure();
 
@@ -64,7 +67,7 @@ Hooks.once("ready", async function () {
   // Notify
   setupNotifySpellstrikeRecharge(getSetting("notify.spellstrike.recharge"));
 
-  setupStartOfSession(getSetting("notify.start-session.journal"))
+  setupStartOfSession(getSetting("notify.start-session.journal"));
 
   // Replace
   setupPauseReplacement();